requirements.txt
asgiref==3.7.2
astroid==2.15.5
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.6
Django==4.2.1
django-extensions==3.2.1
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
flake8==6.0.0
idna==3.4
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
Markdown==3.4.3
MarkupSafe==2.1.3
mccabe==0.7.0
packaging==23.1
Pillow==9.5.0
platformdirs==3.5.1
psycopg2==2.9.6
pycodestyle==2.10.0
pydot==1.4.2
pyflakes==3.0.1
PyJWT==2.7.0
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pyparsing==3.0.9
pytz==2023.3
PyYAML==6.0
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.6.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
wrapt==1.15.0
